requirements.txt
aiogram>=3.17.0
aiohttp>=3.11.11
beautifulsoup4>=4.12.3
pdfplumber>=0.11.5
apscheduler>=3.11.0
requests>=2.32.3
aiosqlite>=0.19.0
python-dotenv>=1.0.0
aiofiles>=24.1.0